moved zod schema to models
